restructuring the traits as we have generalized them
