+++ applications/numeric/doubledouble.cpp
@@ -9,8 +9,6 @@
 #if (__cplusplus == 202003L) || (_MSVC_LANG == 202003L)
 #include <numbers>    // high-precision numbers
 #endif
-#include <universal/common/arithmetic_traits.hpp>
-#include <universal/common/number_traits.hpp>
 #include <universal/verification/performance_runner.hpp>
 #include <universal/verification/test_suite.hpp>
