+++ applications/numeric/numbers.cpp
@@ -4,11 +4,14 @@
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <universal/utility/directives.hpp>
+#include <cstddef>
+#include <stdexcept>
+#include <cstring>
+#include <ostream>
 #include <limits>
 #if (__cplusplus == 202003L) || (_MSVC_LANG == 202003L)
 #include <numbers>    // high-precision numbers
 #endif
-#include <universal/common/arithmetic_traits.hpp>
 
 // select the number systems we would like to compare
 #include <universal/number/integer/integer.hpp>
@@ -17,11 +20,6 @@
 #include <universal/number/cfloat/cfloat.hpp>
 #include <universal/number/posit/posit.hpp>
 #include <universal/number/lns/lns.hpp>
-
-#include <cstddef>
-#include <stdexcept>
-#include <cstring>
-#include <ostream>
 
 #ifndef _MSC_VER
 #  if __cplusplus < 201103
