+++ include/universal/number/bfloat/bfloat.hpp
@@ -7,6 +7,19 @@
 #define _BFLOAT_STANDARD_HEADER_
 
 ////////////////////////////////////////////////////////////////////////////////////////
+///  COMPILATION DIRECTIVES TO DIFFERENT COMPILERS
+
+// compiler specific configuration for long double support
+#include <universal/utility/long_double.hpp>
+// compiler specific configuration for C++20 bit_cast
+#include <universal/utility/bit_cast.hpp>
+
+////////////////////////////////////////////////////////////////////////////////////////
+/// required std libraries 
+#include <iostream>
+#include <iomanip>
+
+////////////////////////////////////////////////////////////////////////////////////////
 ///  BEHAVIORAL COMPILATION SWITCHES
 
 ////////////////////////////////////////////////////////////////////////////////////////
@@ -23,6 +36,12 @@
 // default is to use std::cerr for signalling an error
 #define BFLOAT_THROW_ARITHMETIC_EXCEPTION 0
 #endif
+
+///////////////////////////////////////////////////////////////////////////////////////
+// bring in the trait functions
+#include <universal/traits/number_traits.hpp>
+#include <universal/traits/arithmetic_traits.hpp>
+#include <universal/common/number_traits_reports.hpp>
 
 ////////////////////////////////////////////////////////////////////////////////////////
 /// INCLUDE FILES that make up the library
