+++ include/universal/number/cfloat/cfloat.hpp
@@ -44,6 +44,12 @@
 #define CFLOAT_NATIVE_SQRT 0
 #endif
 
+///////////////////////////////////////////////////////////////////////////////////////
+// bring in the trait functions
+#include <universal/traits/number_traits.hpp>
+#include <universal/traits/arithmetic_traits.hpp>
+#include <universal/common/number_traits_reports.hpp>
+
 ////////////////////////////////////////////////////////////////////////////////////////
 /// INCLUDE FILES that make up the library
 #include <universal/number/cfloat/exceptions.hpp>
