+++ include/universal/number/lns/lns.hpp
@@ -37,6 +37,12 @@
 #define LNS_THROW_ARITHMETIC_EXCEPTION 0
 #endif
 
+///////////////////////////////////////////////////////////////////////////////////////
+// bring in the trait functions
+#include <universal/traits/number_traits.hpp>
+#include <universal/traits/arithmetic_traits.hpp>
+#include <universal/common/number_traits_reports.hpp>
+
 ////////////////////////////////////////////////////////////////////////////////////////
 /// INCLUDE FILES that make up the library
 #include <universal/number/lns/exceptions.hpp>
