+++ static/cfloat/api/traits.cpp
@@ -18,11 +18,6 @@
 #include <universal/number/cfloat/cfloat.hpp>
 #include <universal/verification/test_reporters.hpp>
 
-// bring in the trait functions
-#include <universal/common/number_traits.hpp>
-#include <universal/common/arithmetic_traits.hpp>
-#include <universal/common/number_traits.hpp>
-
 int main()
 try {
 	using namespace sw::universal;
